requirements.txt
astro-pi==1.1.5
boto==2.38.0
Cap1xxx==0.1.2
chardet==2.3.0
codebug-tether==0.9.1
configobj==5.0.6
deja-dup-caja==0.0.4
drumhat==0.0.5
duplicity==0.7.6
envirophat==0.0.6
ExplorerHAT==0.4.2
folder-color-caja==0.0.79
folder-color-common==0.0.79
gpiozero==1.3.1
lockfile==0.12.2
mate-menu==5.7.1
mcpi==0.1.1
microdotphat==0.1.3
mote==0.0.3
motephat==0.0.2
netifaces==0.10.4
numpy==1.11.0
pantilthat==0.0.2
pianohat==0.0.5
picamera==1.12
picraft==1.0
piglow==1.2.4
pigpio==1.30
Pillow==3.1.2
pycrypto==2.6.1
pygame===1.9.1release
pygobject==3.20.0
pyserial==3.0.1
python-cloudfiles==1.7.10
python-distutils-extra==2.39
python-xlib==0.14
pyxdg==0.25
rainbowhat==0.0.1
requests==2.9.1
RPi.GPIO==0.6.3
RTIMULib==7.2.1
scrollphat==0.0.7
sense-emu==1.0
sense-hat==2.2.0
six==1.10.0
sn3218==1.2.6
spidev==3.0
st7036==1.4.3
urllib3==1.13.1
wxPython==3.0.2.0
wxPython-common==3.0.2.0
youtube-dl==2016.2.22
Youtube-DLG==0.3.8
certifi==2018.10.15
chardet==3.0.4
idna==2.7
requests==2.20.0
tuyapy==0.1.3
urllib3==1.24